Improved support for more input types of diffs.
Created one list all_tokens in tokens.cpp, where is stored list of all
tokens and functions which are able to instance them. Drawback of this
solution is that it's necessary to define static function for each diff
type, but this function is very simple, so I decided for this solution.